[agw] Create separate test suite for agw conformance tests (#60037)
* [agw] Create seperate test suite for agw conformance tests


fix incorrect gateway class name for agw tests


Added missing build tag


Removed ambient tests, made the agentgateway skip without explicitly
enabling them, and added a make target to just run the agentgateway
conf tests.


* Make flag and target generic for agw


---------
