+++ pkg/test/framework/resource/flags.go
@@ -244,6 +244,8 @@ func init() {
 	flag.BoolVar(&settingsFromCommandLine.NativeNftables, "istio.test.nativeNftables", settingsFromCommandLine.NativeNftables,
 		"If set, native nftable rules will be used instead of iptable rules for traffic redirection.")
 
+	flag.BoolVar(&settingsFromCommandLine.Agentgateway, "istio.test.agentgateway", settingsFromCommandLine.Agentgateway,
+		"If set, agentgateway conformance tests will be run.")
 	initGatewayConformanceTimeouts()
 }
