+++ pkg/test/framework/resource/settings.go
@@ -202,6 +202,10 @@ type Settings struct {
 
 	// If enabled, native nftable rules will be used for traffic redirection instead of iptable rules.
 	NativeNftables bool
+
+	// Agentgateway indicates that the agentgateway tests should be run. This is gated behind a separate flag since
+	// agentgateway support is experimental.
+	Agentgateway bool
 }
 
 // SkipVMs changes the skip settings at runtime
