+++ tests/integration/tests.mk
@@ -126,5 +126,10 @@ test.integration.kube.environment: | $(JUNIT_REPORT) check-go-tag
 ifeq (${JOB_TYPE},postsubmit)
 	$(call run-test,./tests/integration/...)
 else
-	$(call run-test,./tests/integration/security/ ./tests/integration/pilot,-run="TestReachability|TestTraffic|TestGatewayConformance")
+	$(call run-test,./tests/integration/security/ ./tests/integration/pilot/,-run="TestReachability|TestTraffic|TestGatewayConformance")
 endif
+
+# Agentgateway support is currently experimental. Only used to run agentgateway tests as optional
+.PHONY: test.integration.kube.agentgateway
+test.integration.kube.agentgateway: | $(JUNIT_REPORT) check-go-tag
+	$(call run-test,./tests/integration/pilot/agentgateway/)
